requirements.txt
pillow
opencv-python
torch 
torchvision
numpy
matplotlib
tensorboardX